refactor: #58 Repackaged statistics-related classes into one "statistics" package. Draft of Config class added.
