+++ .run/PlayerDataDownloader.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="PlayerDataDownloader" type="Application" factoryName="Application">
-    <option name="MAIN_CLASS_NAME" value="ru.klavogonki.kgparser.PlayerDataDownloader" />
+    <option name="MAIN_CLASS_NAME" value="ru.klavogonki.kgparser.statistics.download.PlayerDataDownloader" />
     <module name="kgparser-srv" />
     <option name="PROGRAM_PARAMETERS" value="$PROJECT_DIR$/../kg/ 100000 110000 100" />
     <extension name="coverage">
