+++ .run/PlayerJsonParser.run.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="PlayerJsonParser" type="Application" factoryName="Application" nameIsGenerated="true">
-    <option name="MAIN_CLASS_NAME" value="ru.klavogonki.kgparser.PlayerJsonParser" />
+    <option name="MAIN_CLASS_NAME" value="ru.klavogonki.kgparser.statistics.download.PlayerJsonParser" />
     <module name="kgparser-srv" />
     <option name="PROGRAM_PARAMETERS" value="C:\java\kg\home\ec2-user\kg 1 625000 &quot;2020-12-08 02-39-07&quot;" />
     <extension name="coverage">
