+++ kgparserSpringBoot/src/main/java/ru/klavogonki/kgparser/springboot/KgParserApplication.java
@@ -9,9 +9,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import ru.klavogonki.kgparser.PlayerDataDownloader;
-import ru.klavogonki.kgparser.PlayerJsonData;
-import ru.klavogonki.kgparser.PlayerJsonParser;
 import ru.klavogonki.kgparser.export.ExportContext;
 import ru.klavogonki.kgparser.export.IndexPageExporter;
 import ru.klavogonki.kgparser.export.PlayersByRankExporter;
@@ -40,6 +37,9 @@ import ru.klavogonki.kgparser.jsonParser.mapper.PlayerMapper;
 import ru.klavogonki.kgparser.jsonParser.mapper.PlayerVocabularyStatsMapper;
 import ru.klavogonki.kgparser.jsonParser.repository.PlayerRepository;
 import ru.klavogonki.kgparser.jsonParser.repository.PlayerVocabularyStatsRepository;
+import ru.klavogonki.kgparser.statistics.download.PlayerDataDownloader;
+import ru.klavogonki.kgparser.statistics.download.PlayerJsonData;
+import ru.klavogonki.kgparser.statistics.download.PlayerJsonParser;
 import ru.klavogonki.kgparser.util.DateUtils;
 import ru.klavogonki.openapi.model.GetStatsOverviewGameType;
 import ru.klavogonki.openapi.model.GetStatsOverviewResponse;
