+++ kgparserSpringBoot/src/test/java/ru/klavogonki/kgparser/jsonParser/mapper/PlayerMapperTest.java
@@ -4,11 +4,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mapstruct.factory.Mappers;
 import ru.klavogonki.kgparser.Car;
-import ru.klavogonki.kgparser.PlayerJsonData;
 import ru.klavogonki.kgparser.Rank;
-import ru.klavogonki.kgparser.jsonParser.ApiErrors;
 import ru.klavogonki.kgparser.jsonParser.Assertions;
 import ru.klavogonki.kgparser.jsonParser.entity.PlayerEntity;
+import ru.klavogonki.kgparser.statistics.download.ApiErrors;
+import ru.klavogonki.kgparser.statistics.download.PlayerJsonData;
 import ru.klavogonki.kgparser.util.DateUtils;
 import ru.klavogonki.openapi.model.Bio;
 import ru.klavogonki.openapi.model.GetIndexDataResponse;
