+++ kgparserSrv/src/main/java/ru/klavogonki/kgparser/statistics/download/ApiErrors.java
@@ -1,4 +1,4 @@
-package ru.klavogonki.kgparser.jsonParser;
+package ru.klavogonki.kgparser.statistics.download;
 
 public final class ApiErrors {
