+++ kgparserSrv/src/main/java/ru/klavogonki/kgparser/statistics/download/PlayerJsonParser.java
@@ -1,11 +1,15 @@
-package ru.klavogonki.kgparser;
+package ru.klavogonki.kgparser.statistics.download;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import ru.klavogonki.kgparser.jsonParser.ApiErrors;
-import ru.klavogonki.kgparser.jsonParser.JacksonUtils;
+import ru.klavogonki.kgparser.Car;
+import ru.klavogonki.kgparser.Dictionary;
+import ru.klavogonki.kgparser.DictionaryMode;
+import ru.klavogonki.kgparser.Rank;
+import ru.klavogonki.kgparser.StandardDictionary;
 import ru.klavogonki.kgparser.util.DateUtils;
+import ru.klavogonki.kgparser.util.JacksonUtils;
 import ru.klavogonki.openapi.model.GetIndexDataResponse;
 import ru.klavogonki.openapi.model.GetIndexDataStats;
 import ru.klavogonki.openapi.model.GetStatsOverviewGameType;
