+++ kgparserSrv/src/main/resources/statistics/config.json
@@ -1,0 +1,7 @@
+{
+  "jsonFilesRootDir": "D:/kg/json",
+  "threadsCount": 100,
+  "minPlayerId": 1,
+  "maxPlayerId": 628000,
+  "statisticsPagesRootDir": "C:/java/kgparser/kgstats/src/main/webapp/"
+}
