+++ kgparserSrv/src/test/java/ru/klavogonki/kgparser/statistics/download/PlayerDataDownloaderTest.java
@@ -1,4 +1,4 @@
-package ru.klavogonki.kgparser;
+package ru.klavogonki.kgparser.statistics.download;
 
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.junit.jupiter.api.Test;
