+++ kgparserSrv/src/test/java/ru/klavogonki/kgparser/statistics/download/PlayerJsonParserTest.java
@@ -1,8 +1,8 @@
-package ru.klavogonki.kgparser;
+package ru.klavogonki.kgparser.statistics.download;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import ru.klavogonki.kgparser.jsonParser.ApiErrors;
+import ru.klavogonki.kgparser.StandardDictionary;
 import ru.klavogonki.kgparser.util.TestUtils;
 import ru.klavogonki.openapi.model.Bio;
 import ru.klavogonki.openapi.model.BioAssert;
